+++ src/main/java/ru/vyarus/guice/persist/orient/repository/command/ext/listen/live/LiveListenerTypeSupport.java
@@ -10,9 +10,9 @@ import com.orientechnologies.orient.core.sql.query.OLiveQuery;
 import com.orientechnologies.orient.core.sql.query.OLiveResultListener;
 import ru.vyarus.guice.persist.orient.db.PersistentContext;
 import ru.vyarus.guice.persist.orient.repository.command.ext.listen.Listen;
-import ru.vyarus.guice.persist.orient.repository.command.ext.listen.ListenerTypeSupport;
+import ru.vyarus.guice.persist.orient.repository.command.ext.listen.support.ListenerTypeSupport;
 import ru.vyarus.guice.persist.orient.repository.command.live.LiveQuery;
-import ru.vyarus.guice.persist.orient.repository.command.live.mapper.LiveResultListener;
+import ru.vyarus.guice.persist.orient.repository.command.live.mapper.LiveQueryListener;
 import ru.vyarus.guice.persist.orient.repository.command.live.mapper.LiveResultMapper;
 import ru.vyarus.guice.persist.orient.repository.core.ext.service.result.converter.RecordConverter;
 import ru.vyarus.guice.persist.orient.repository.core.spi.parameter.ParamInfo;
@@ -23,7 +23,7 @@ import static ru.vyarus.guice.persist.orient.repository.core.MethodExecutionExce
 /**
  * Handler for {@link Listen} parameters within {@link LiveQuery}.
  * <p>
- * Listener may be {@link OLiveResultListener} (orient) or {@link LiveResultListener} if special result
+ * Listener may be {@link OLiveResultListener} (orient) or {@link LiveQueryListener} if special result
  * conversions required.
  * <p>
  * If transaction is required (through {@link Listen}) then listener will be wrapped to apply transaction.
@@ -45,28 +45,30 @@ public class LiveListenerTypeSupport implements ListenerTypeSupport {
         check(int.class.equals(returnType) || Integer.class.equals(returnType),
                 "Live query method must have int return type to receive subscription token");
         check(OLiveResultListener.class.isAssignableFrom(param.type)
-                        || LiveResultListener.class.isAssignableFrom(param.type),
+                        || LiveQueryListener.class.isAssignableFrom(param.type),
                 "Only %s or %s can be used as live listener",
-                OLiveResultListener.class.getName(), LiveResultListener.class.getName());
+                OLiveResultListener.class.getName(), LiveQueryListener.class.getName());
     }
 
     @Override
     public OCommandResultListener processListener(final OCommandRequest command,
                                                   final Object listener,
                                                   final Injector injector,
-                                                  final boolean transactional) {
+                                                  final boolean transactional,
+                                                  final Class<?> conversionTarget) {
         checkExec(command instanceof OLiveQuery,
                 "Live listener (@%s parameter) can only be used with @%s",
                 Listen.class.getSimpleName(), LiveQuery.class.getSimpleName());
-        return wrap(listener, injector, transactional);
+        return wrap(listener, injector, transactional, conversionTarget);
     }
 
     private OCommandResultListener wrap(final Object listener,
                                         final Injector injector,
-                                        final boolean transactional) {
-        final OLiveResultListener adaptedListener = listener instanceof LiveResultListener
+                                        final boolean transactional,
+                                        final Class<?> targetType) {
+        final OLiveResultListener adaptedListener = listener instanceof LiveQueryListener
                 // special listener with custom mapping
-                ? wrap((LiveResultListener) listener, injector) : (OLiveResultListener) listener;
+                ? wrap((LiveQueryListener) listener, injector, targetType) : (OLiveResultListener) listener;
 
         final OCommandResultListener res;
         if (transactional) {
@@ -82,9 +84,11 @@ public class LiveListenerTypeSupport implements ListenerTypeSupport {
         return res;
     }
 
-    private OLiveResultListener wrap(final LiveResultListener listener, final Injector injector) {
+    private OLiveResultListener wrap(final LiveQueryListener listener,
+                                     final Injector injector,
+                                     final Class<?> targetType) {
         final RecordConverter converter = injector.getInstance(RecordConverter.class);
-        return new LiveResultMapper(converter, listener);
+        return new LiveResultMapper(converter, listener, targetType);
     }
 
 }