+++ src/main/java/ru/vyarus/guice/persist/orient/repository/command/live/LiveQuery.java
@@ -15,9 +15,9 @@ import java.lang.annotation.Target;
  * with {@link ru.vyarus.guice.persist.orient.repository.command.ext.listen.Listen}.
  * Method must return int or Integer (subscription token required to unsubscribe).
  * <p>
- * {@link ru.vyarus.guice.persist.orient.repository.command.live.mapper.LiveResultListener} could be used as
+ * {@link ru.vyarus.guice.persist.orient.repository.command.live.mapper.LiveQueryListener} could be used as
  * listener in order to apply automatic conversions (like in usual repository methods). For example:
- * {@code @LiveQuery("select from Model") int subscribe(@Listen LiveResultListener<Model> listener)}.
+ * {@code @LiveQuery("select from Model") int subscribe(@Listen LiveQueryListener<Model> listener)}.
  * <p>
  * By default listener is executed within transaction (see
  * {@link ru.vyarus.guice.persist.orient.repository.command.ext.listen.Listen}).
@@ -32,7 +32,7 @@ import java.lang.annotation.Target;
  * @see <a href="http://orientdb.com/docs/last/Live-Query.html">docs</a>
  * @see ru.vyarus.guice.persist.orient.repository.command.ext.listen.Listen
  * @see com.orientechnologies.orient.core.sql.query.OLiveResultListener
- * @see ru.vyarus.guice.persist.orient.repository.command.live.mapper.LiveResultListener
+ * @see ru.vyarus.guice.persist.orient.repository.command.live.mapper.LiveQueryListener
  * @since 29.09.2017
  */
 @Target(ElementType.METHOD)