+++ CHANGELOG.md
@@ -4,10 +4,13 @@
 * (breaking) @Listen parameter now wraps provided listener with transaction by default (may be disabled: transactional = false)
 * Add @LiveQuery repository extension (live queries support)
     - Listener parameter recognized by @Listen annotation
-    - custom LiveResultListener may be used to automatically convert record to object or vertex
-* New service RecordConverter may be used directly to convert ODocument to object or vertex and apply projection like in repositories methods
+    - custom LiveQueryListener interface may be used to automatically convert record to object or vertex
+* New service RecordConverter may be used directly to:
+    - convert ODocument to object or vertex 
+    - apply default repository method converter (e.g. apply projection)
 (when document contains just one value it can be converted to value directly).    
-    
+* Add QueryListener interface to use instead of OCommandResultListener for @Query and @AsyncQuery (under @Listen)
+which performs automatic result conversion (with RecordConverter to mimic the same behaviour as for repository return).
 
 ### 3.2.0 (2016-09-25)
 * Update to orient 2.1